- Fixes some errors related to arrays of records containing file.     - Add logic to handle arrays of files and arrays of records containing files.
This change breaks arrays of records as areguments to read[ln] and write[ln].  For the case arguments to these functions that are arrays of of RECORDs, it is not possible to know if the first argument is a file number or not.  We have to parse the field selector to know that and, being a single pass compiler, this is tricky.

This change breaks test case 011